requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
turbopuffer==0.1.0
SpeechRecognition==3.10.0
pydub==0.25.1
pyaudio==0.2.13 